backend/requirements.txt
anthropic==0.94.0
beautifulsoup4==4.14.3
fastapi==0.135.3
pymongo==4.16.0
python-dotenv==1.1.0
requests==2.32.5
uvicorn==0.44.0
